Configurations

In Skylight.jl, the configurations objects reunite the main components of the astrophysical problems: the spacetime, the radiative model, and the camera. Additionally, it contains a parameter to completely determine the geometrized code units together with the $G = c = 1$ conditions. There are different configuration types depending on the method (observer-to-emitter or emitter-to-observer) and whether the radiative transport occurs in a vacuum, which allows for certain simplifications. However, the emitter-to-observer method is not fully implemented yet in this version of Skylight.jl, so we only consider the observer-to-emitter method.

Examples

For vacuum transport problems using the observer-to-emitter method, the configurations are implemented by the VacuumOTEConfigurations type. They can be constructed as follows:

configurations = VacuumOTEConfigurations(spacetime = spacetime,
    camera = camera,
    radiative_model = model,
    unit_mass_in_solar_masses = 1.0)

where spacetime is the spacetime object, camera is the camera object, and model is the radiative model object. The unit_mass_in_solar_masses is equal to $M/M_\odot$ where $M=1$ is the unit mass of the geometrized unit system of the code. An important point is that this parameter must equal the value set in any radiative model object that requires it, as the NovikovThorneDisk model for example. We acknowledge this is a non-ideal solution, and we are working on a more consistent way to handle the units in Skylight.jl. The advantage of this method is that it avoids re-running ray-tracing when the absolute unit scale changes.

For non-vacuum transport problems, such as those involving an IonTorus radiative model, the configurations are implemented by the NonVacuumOTEConfigurations type. The can be constructed as follows:

configurations = NonVacuumOTEConfigurations(spacetime = spacetime,
    camera = camera,
    radiative_model = model,
    unit_mass_in_solar_masses = 1.0,
    observation_energies = [1e-8, 1e-7])

The observation_energies parameter specifies the (CGS) energies at which the specific intensity is to be computed in the observation frame. In non-vacuum problems, this has to be specified right away, as the solution depends on the energy in a non-trivial manner, as opposed to vacuum problems where only the invariant $I_\nu / \nu^3$ needs to be connected between the observation and emission frames.

Types

Skylight.VacuumOTEConfigurations — Type

VacuumOTEConfigurations{S,M,C} <: AbstractOTEConfigurations

Cconfigurations object for vacuum observer-to-emitter (OTE) radiative transport problems.

Fields

spacetime::S: The spacetime in which the transport problem is set. Must be a subtype of AbstractSpacetime.

radiative_model::M: The radiative model describing the source of radiation. Must be a subtype of AbstractRadiativeModel and non-vacuum.

camera::C: The observational setup, including position and orientation. Must be a subtype of AbstractCamera.

unit_mass_in_solar_masses::Float64: The unit mass in solar masses used to determine the geometrized code units. Must be positive.

Constructor

configurations = NonVacuumOTEConfigurations(
    spacetime = spacetime,
    radiative_model = model,
    camera = camera,
    unit_mass_in_solar_masses = 1.0)

Skylight.NonVacuumOTEConfigurations — Type

NonVacuumOTEConfigurations{S,M,C} <: AbstractOTEConfigurations

Cconfigurations object for non-vacuum observer-to-emitter (OTE) radiative transport problems.

Fields

spacetime::S: The spacetime in which the transport problem is set. Must be a subtype of AbstractSpacetime.

radiative_model::M: The radiative model describing the source of radiation. Must be a subtype of AbstractRadiativeModel and non-vacuum.

camera::C: The observational setup, including position and orientation. Must be a subtype of AbstractCamera.

observation_energies::Vector{Float64}: A vector of energies (in ergs) at which the specific intensity is to be computed. All energies must be positive.

unit_mass_in_solar_masses::Float64: The unit mass in solar masses used to determine the geometrized code units. Must be positive.

Constructor

configurations = NonVacuumOTEConfigurations(
    spacetime = spacetime,
    radiative_model = model,
    camera = camera,
    observation_energies = [1e-8, 1e-7],
    unit_mass_in_solar_masses = 1.0)