Catalogue of radiative models

Geometrically thin, optically thick accretion disks

Shakura & Sunyaev accretion disk

Skylight.ShakuraSunyaevDisk — Type

ShakuraSunyaevDisk <: AbstractAccretionDisk

Shakura & Sunyaev geometrically thin, optically thick accretion disk model.

Fields

inner_radius::Float64: The inner of the accretion disk. Must be larger than or equal to zero.

outer_radius::Float64: The outer radius of the accretion disk. Must be larger than or equal to inner_radius.

M1::Float64: The unitary mass in solar masses. Must be positive.

Mdot_to_MEdd::Float64: The accretion rate in units of the Eddington accretion rate. Must be positive.

η::Float64: The radiative efficiency of the disk, which must be in the range (0, 1].

rotation_sense::AbstractRotationSense: The sense of rotation of the disk, which can be either ProgradeRotation() or RetrogradeRotation(). Default is ProgradeRotation().

Examples

spacetime = KerrSpacetimeBoyerLindquistCoordinates(M=1.0, a=0.5)
risco = innermost_circular_orbit_radius(spacetime, ProgradeRotation())
disk = ShakuraSunyaevDisk(inner_radius = risco, outer_radius=1000.0, M1=1e7, Mdot_to_MEdd=0.1, η=0.1)

Novikov & Thorne accretion disk

Skylight.NovikovThorneDisk — Type

NovikovThorneDisk <: AbstractAccretionDisk

Novikov & Thorne geometrically thin, optically thick accretion disk model around a Schwarzschild/Kerr black hole.

Fields

inner_radius::Float64: The inner radius of the accretion disk. Must be larger than or equal to zero.

outer_radius::Float64: The outer radius of the accretion disk. Must be larger than or equal to inner_radius.

M1::Float64: The unitary mass in solar masses. Must be positive.

Mdot_to_MEdd::Float64: The accretion rate in units of the Eddington accretion rate. Must be positive.

η::Float64: The radiative efficiency of the disk, which must be in the range (0, 1].

rotation_sense::AbstractRotationSense: The sense of rotation of the disk, which can be either ProgradeRotation() or RetrogradeRotation(). Default is ProgradeRotation().

Examples

spacetime = KerrSpacetimeBoyerLindquistCoordinates(M=1.0, a=0.5)
risco = innermost_circular_orbit_radius(spacetime, ProgradeRotation())
disk = NovikovThorneDisk(inner_radius = risco, outer_radius=1000.0, M1=1e7, Mdot_to_MEdd=0.1, η=0.1)

RAR extension of Shakura & Sunyaev accretion disk

Skylight.RARDisk — Type

RARDisk <: AbstractAccretionDisk

RAR extension of the Shakura & Sunyaev geometrically thin, optically thick accretion disk model. See Millauro et al. (2024).

Fields

inner_radius::Float64: The inner of the accretion disk. Must be larger than or equal to zero.

outer_radius::Float64: The outer radius of the accretion disk. Must be larger than or equal to inner_radius.

M1::Float64: The unitary mass in solar masses. Must be positive.

Mdot_to_MEdd::Float64: The accretion rate in units of the Eddington accretion rate. Must be positive.

η::Float64: The radiative efficiency of the disk, which must be in the range (0, 1].

rotation_sense::AbstractRotationSense: The sense of rotation of the disk, which can be either ProgradeRotation() or RetrogradeRotation(). Default is ProgradeRotation().

Examples

disk = RARDisk(inner_radius=0.0, outer_radius=1000.0, M1=1e7, Mdot_to_MEdd=0.1, η=0.1)

Tabulated temperature

Skylight.AccretionDiskWithTabulatedTemperature — Type

AccretionDiskWithTabulatedTemperature <: AbstractAccretionDisk

Geometrically thin, optically thick accretion disk model with tabulated temperature.

Fields

inner_radius::Float64: The inner of the accretion disk. Must be larger than or equal to zero.

outer_radius::Float64: The outer radius of the accretion disk. Must be larger than or equal to inner_radius.

rotation_sense::AbstractRotationSense: The sense of rotation of the disk, which can be either ProgradeRotation() or RetrogradeRotation(). Default is ProgradeRotation().

filename::String: The name of a two-column file containing the temperature profile as radius vs. temperature.

Examples

disk = AccretionDiskWithTabulatedTemperature(inner_radius=6.0, outer_radius=1000.0, filename="temperature.txt")

Flat spacetime lamppost corona line emission profile

Skylight.AccretionDiskWithFlatLamppostProfile — Type

AccretionDiskWithFlatLamppostProfile <: AbstractAccretionDisk

Geometrically thin, optically thick accretion disk model with lamppost corona line emissivity profile as in flat spacetime:

$\epsilon \propto \frac{h}{(r^2+h^2)^{3/2}}$

where h is the height of the corona above the disk plane.

Fields

inner_radius::Float64: The inner of the accretion disk. Must be larger than or equal to zero.

outer_radius::Float64: The outer radius of the accretion disk. Must be larger than or equal to inner_radius.

rotation_sense::AbstractRotationSense: The sense of rotation of the disk, which can be either ProgradeRotation() or RetrogradeRotation(). Default is ProgradeRotation().

corona_height::String: The height of the corona above the disk plane.

Examples

disk = AccretionDiskWithFlatLamppostProfile(inner_radius=6.0, outer_radius=1000.0, corona_height=5.0)

Tabulated line emission profile

Skylight.AccretionDiskWithTabulatedProfile — Type

AccretionDiskWithTabulatedProfile <: AbstractAccretionDisk

Geometrically thin, optically thick accretion disk model with tabulated line emission radial profile.

Fields

inner_radius::Float64: The inner of the accretion disk. Must be larger than or equal to zero.

outer_radius::Float64: The outer radius of the accretion disk. Must be larger than or equal to inner_radius.

rotation_sense::AbstractRotationSense: The sense of rotation of the disk, which can be either ProgradeRotation() or RetrogradeRotation(). Default is ProgradeRotation().

filename::String: The name of a two-column file containing the line emission radial profile as radius vs. profile.

Examples

disk = AccretionDiskWithTabulatedProfile(inner_radius=6.0, outer_radius=1000.0, filename="emissivity.txt")

Geometrically thick, optically thin accretion disks

Ion torus

Skylight.IonTorus — Type

IonTorus <: AbstractRadiativeModel

Geometrically thick, optically thin ion torus model with synchrotron and bremsstahlung emission (Straub et al. 2012).

Fields

λ::Float64: Specific angular momentum dimensionless parameter

ϵc::Float64: Central energy density in CGS

n::Float64: Polytropic index

Tec::Float64: Central electron temperature in Kelvin

ξ::Float64: Electron to proton temperature ratio at the center

β::Float64: Equipartition factor

H_abundance::Float64: Hydrogen abundance

He_abundance::Float64: Helium abundance

rotation_sense::R: Sense of rotation of the torus, which can be either ProgradeRotation() or RetrogradeRotation()

radiative_process::P: Radiative process, which can be either Bremsstrahlung(), Synchrotron() or SynchrotronAndBremsstrahlung()

Constructor

spacetime = KerrSpacetimeBoyerLindquistCoordinates(M = 1.0, a = 0.7)
IonTorus = IonTorus(spacetime; 
    λ = 0.3, 
    ϵc = 1e-17, 
    n = 3 / 2, 
    Tec = 2e9, 
    ξ = 0.1, 
    β = 0.1, 
    H_abundance = 0.75, 
    He_abundance = 0.25, 
    rotation_sense = ProgradeRotation(), 
    radiative_process = Bremsstrahlung())

Note

Synchrotron self-absorption is not implemented yet.

Others

Lamppost corona

Skylight.LamppostCorona — Type

LamppostCorona <: AbstractCorona

Lamppost corona model as described in Dauser et al. (2013).

Fields

height::Float64: The height of the lamppost corona.

spectral_index::Float64: The spectral index of the power-law photon emission. Default is 2.0.

theta_offset::Float64: The offset of the lamppost corona in the polar angle. Default is 1e-8. Must be small but nonzero to avoid the coordinate singularity at the polar axis.

Examples

corona = LamppostCorona(height=2.5, spectral_index = 2.0)

Synthetic polar cap

Skylight.CircularHotSpot — Type

CircularHotSpot <: AbstractSurfaceEmissionModel

Circular hot spot on the surface of a neutron star.

Fields

star_radius_in_km::Float64: The radius of the star in km. Must be positive.

angular_speed_in_Hz::Float64: The angular speed of the star in Hz. Must be non-zero.

center_colatitude_in_degrees::Float64: The colatitude of the spot center in degrees. Must be in the range [0, 90].

angular_radius_in_radians::Float64: The angular radius of the polar cap, in radians. Must be positive.

temperature::Float64: The temperature of the polar cap. Must be positive.

Examples

hot_spot = CircularHotSpot(
    star_radius_in_km = 12.0,
    angular_speed_in_Hz = 200,
    center_colatitude_in_degrees = 30.0,
    angular_radius_in_radians = 1.0,
    temperature_in_keV = 0.35
    M1 = 1.4
)

Star across traversable wormhole

Skylight.StarAcrossWormhole — Type

StarAcrossWormhole <: AbstractSurfaceEmissionModel

Uniform temperature star accross a traversable wormhole. See also (see ChargedWormholeSpacetimeRegularCoordinates).

Fields

l_center::Float64: The center of the star in the regular radial coordinate.

star_radius::Float64: The radius of the star.

Examples

star = StarAcrossWormhole(l_center=5.0, star_radius=1.0)