Functions

Generic

Skylight.rest_frame_four_velocity! — Method

rest_frame_four_velocity!(vector::AbstractVector, position::AbstractVector, metric::AbstractMatrix, spacetime::AbstractSpacetime, model::AbstractRadiativeModel, coords_top::AbstractCoordinatesTopology, spacetime_cache::AbstractSpacetimeCache, model_cache::AbstractModelCache)

Rest frame four velocity of the model at given `position`. This is the frame where the model radiative functions are defined.

#Arguments

vector: Output vector.

position: Position where the four-velocity is evaluated.

metric: Metric tensor at position.

spacetime: Spacetime.

model: Radiative model.

coords_top: Coordinates topology.

spacetime_cache: Spacetime cache.

model_cache: Model cache.

See also

allocate_cache(model::AbstractRadiativeModel)

Skylight.rest_frame_bolometric_intensity — Method

rest_frame_bolometric_intensity(position::AbstractVector, momentum::AbstractVector, rest_frame_four_velocity::AbstractVector, metric::AbstractMatrix, spacetime::AbstractSpacetime, model::AbstractRadiativeModel, coords_top::AbstractCoordinatesTopology, cache::AbstractModelCache)

Bolometric intensity of the model at given `position`.

#Arguments

position: Position where the intensity is evaluated.

momentum: Momentum of the emission (frequency and direction).

rest_frame_four_velocity: Rest frame four velocity of the model at position. Must but be normalized.

metric: Metric tensor at position.

spacetime: Spacetime.

model: Radiative model.

coords_top: Coordinates topology.

cache: Model cache.

See also

allocate_cache(model::AbstractRadiativeModel)

Skylight.rest_frame_specific_intensity — Method

rest_frame_specific_intensity(position::AbstractVector, momentum::AbstractVector, energy::Real, rest_frame_four_velocity::AbstractVector, metric::AbstractMatrix, spacetime::AbstractSpacetime, model::AbstractRadiativeModel, coords_top::AbstractCoordinatesTopology, cache::AbstractModelCache)

Specific intensity of the model at given `position`.

#Arguments

position: Position where the intensity is evaluated.

momentum: Momentum of the emission (frequency and direction).

energy: Energy of the emission. This should be assumed to be in CGS units.

rest_frame_four_velocity: Rest frame four velocity of the model at position. Must but be normalized.

metric: Metric tensor at position.

spacetime: Spacetime.

model: Radiative model.

coords_top: Coordinates topology.

cache: Model cache.

See also

allocate_cache(model::AbstractRadiativeModel)

Skylight.is_final_position_at_source — Method

is_final_position_at_source(position::AbstractVector, spacetime::AbstractSpacetime, model::AbstractRadiativeModel)

Check if the final position of the photon is at the source of the model. This function is used in the observer-to-emitter method
in vacuum to discard rays that do not intersect the source.

#Arguments

position: Final position of the geodesic.

spacetime: Spacetime.

model: Radiative model.

Skylight.lorentz_factors — Method

lorentz_factors(positions, spacetime::AbstractSpacetime, model::AbstractRadiativeModel)

Lorentz factors of the rest frame four velocities of `model` at given list of positions. Positions should be an iterable
object with the positions as elements.

# Returns

An array with the Lorentz factors of the rest frame four velocities of the model at the given positions.

# Example

```julia
positions = [[0.0, 5.0, 0.0, 0.0], [1.0, 0.0, 5.0, 0.0]]
spacetime = SchwarzschildSpacetimeSphericalCoordinates(M=1.0)
model = DummyModel()
γ = lorentz_factors(positions, spacetime, model)
```

Skylight.allocate_cache — Method

allocate_cache(model::AbstractRadiativeModel)

Allocate a cache object for the given model. The cache object is used to store temporary data in radiative-model-related calculations.

For surface emission models

Skylight.surface_differential! — Method

surface_differential!(differential::AbstractVector, position::AbstractVector, model::AbstractSurfaceEmissionModel, coords_top::AbstractCoordinatesTopology)

Differential of the function defining the emitting surface in the model. For example, for an emitting sphere in Cartesian coordinates, the output would be `[0,2x,2y,2z]`.
The normalization of the differential is not important, as it is only used to calculate the surface (unit) normal in terms of the metric.

#Arguments

differential: Output vector.

position: Position where the differential is evaluated.

model: Radiative model.

coords_top: Coordinates topology.

For thermal emission models

Skylight.temperature — Method

temperature(position::AbstractVector, spacetime::AbstractSpacetime, model::AbstractRadiativeModel)

Temperature of the model at given `position`.

For line emission models

Skylight.line_emission_profile — Method

line_emission_profile(position::AbstractVector, momentum::AbstractVector, rest_frame_four_velocity::AbstractVector, metric::AbstractMatrix, spacetime::AbstractSpacetime, model::AbstractRadiativeModel, coords_top::AbstractCoordinatesTopology, cache::AbstractModelCache)

Emissivity profile for line emission radiative models.

#Arguments

position: Position of the emission.

momentum: Momentum of the emission (frequency and direction). Must be null.

rest_frame_four_velocity: Rest frame four velocity of the model at position. Must but be normalized.

metric: Metric tensor at position.

spacetime: Spacetime.

model: Radiative model.

coords_top: Coordinates topology.

cache: Model cache.

See also

allocate_cache(model::AbstractRadiativeModel)