Skylight.allocate_cache — Method

allocate_cache(spacetime)

Allocate a cache object for the given spacetime. The cache object is used to store temporary data in spacetime-related calculations.

Skylight.allocate_christoffel_cache — Method

allocate_christoffel_cache(spacetime)

Allocate a cache object for the Christoffel symbols calculation. Unless a more specific method is defined, an instance of Skylight.AutoDiffChristoffelCache is returned.

Skylight.approximate_gradient_norm — Method

approximate_gradient_norm(z::Array, dx_spacing::Real, dy_spacing::Real)

Compute the norm of the gradient of a function given a 2D grid of function values z and grid spacings dx_spacing and dy_spacing.

The gradient is computed using central differences in the interior and forward or backward differences at the edges.

Arguments

z::Array: 2D grid of function values.

dx_spacing::Real: Grid spacing in the x direction.

dy_spacing::Real: Grid spacing in the y direction.

Returns

grad_norm::Array: 2D grid of gradient magnitudes.

Skylight.average_inside_bins — Method

average_inside_bins(q, x, y, xbins, ybins)

Calculate the averages of a quantity q at values (x,y) in each bin defined by xbins and ybins.

Skylight.average_inside_bins — Method

average_inside_bins(q, x, bins)

Calculate the averages of a quantity q at value x in each bin defined by bins.

Skylight.axial_symmetry — Method

axial_symmetry(spacetime)

Return IsAxiallySymmetric() if the spacetime is axially symmetric, IsNotAxiallySymmetric() otherwise.

Skylight.bin_values_and_sum_weights — Method

bin_values_and_sum_weights(bins, values, weights)

Bin values and sum weights in each bin.

Arguments

bins: Array of bin edges.

values: Array of values to be binned.

weights: Array of weights to be summed in each bin.

Returns

binned_values: Array of the sum of weights in each bin.

Notes

The length of values and weights must be the same.

Values outside the range of bins are ignored.

Skylight.christoffel! — Method

christoffel!(Γ, position, spacetime, cache)

Evaluate the Christoffel symbols at the given position and store the result in the given array using cache for temporary storage. Unless a more specific method is defined, automatic differentiation is used to compute the Christoffel symbols.

Skylight.christoffel! — Method

" Calculates the Christoffel symbols of a given spacetime metric using the forward-mode automatic differentiation package ForwardDiff.

Arguments:

Γ₂: mutable array of size (4,4,4) to store the resulting Christoffel symbols.

position: tuple of four numbers representing a position in spacetime.

spacetime: object representing the spacetime.

cache: object of type AutoDiffChristoffelCache with containers for metric elements and derivatives.

Returns: nothing.

Note: for the automatic differentiation to work on a given spacetime, any spacetime cache array used in metric calculations must be wrapped by the DiffCache() method as:

```
@with_kw mutable struct KerrSpacetimeCache{T} <: AbstractSpacetimeCache
    l::T = DiffCache(zeros(4))
end
```

This is because automatic differentiation keeps two versions of each variable, a Real and a Dual version, the latter being used to compute derivatives at each node of the chain rule. Also, in the metric!(g, position, spacetime) function, the caches must be accessed via the function get_tmp as in get_tmp(spacetime.l, position), so that the appropriate version of the cache is returned according to the element type of position when metric! is called.

Skylight.circular_geodesic_angular_speed — Method

circular_geodesic_angular_speed(position, spacetime, rotation_sense::AbstractRotationSense)

Angular speed of a circular geodesic at the given position in the given spacetime for a particle rotating in the given sense.

Skylight.circular_geodesic_energy — Method

Set for M=1

Skylight.circular_geodesic_specific_angular_momentum — Method

Set for M=1

Skylight.compose_N_times — Method

compose_N_times(N, f, x)

Compose function f to the input x for N times.

Arguments

N: The number of times to apply the function f.

f: The function to be applied.

x: The input to the function f.

Example

matrix = rand(Bool, 5, 5)
compose_N_times(3, detect_edges, matrix)

Skylight.cos_angle_between_null_vectors — Method

The cosine of the angle between null vectors v and w as seen by observer at normalized four-velocity u

Skylight.cos_angle_between_vectors — Method

The cosine of the angle between vectors v and w as seen by observer at normalized four-velocity

Skylight.det2x2 — Method

Computes the determinant of a 2x2 matrix using a closed-form formula.

Arguments:

matrix: array of size (2,2) containing the input matrix.

Returns: the determinant of the input matrix.

Skylight.det2x2sym — Method

Computes the determinant of a 2x2 symmetric matrix using a closed-form formula.

Arguments:

matrix: array of size (2,2) containing the input symmetric matrix.

Returns: the determinant of the input matrix.

Skylight.det4x4sym — Method

Computes the determinant of a 4x4 symmetric matrix using a closed-form formula.

Arguments:

matrix: array of size (4,4) containing the input symmetric matrix.

Returns: the determinant of the input matrix.

Skylight.detect_edges — Method

detect_edges(N, arr)

Detects values that are at distance at most N from an edge. Composes detect_edges to the input arr for 1 to N times, and combines the results using an elementwise OR operation. Each application of detect_edges is performed on the result of the previous application and the results are combined in succession.

Arguments

N::Int: The maximum number of times to apply the function f.

arr::Union{Matrix{Bool}, BitMatrix}: The boolean array.

Example

julia matrix = rand(Bool, 5, 5) detect_edges(3, matrix)` In this example, the detect_edges function is applied 1 to 3 times to a 5x5 random Boolean matrix, and the results are combined with an OR operation.

Skylight.detect_edges — Method

detect_edges(arr::Array{Bool, 2})

Detects edges in a two-dimensional Boolean array by comparing each element with its neighbors in all four cardinal directions (up, down, left, and right). An edge is considered to be present at a location if the Boolean value at that location is different from the value of any of its neighbors.

Arguments

arr::Array{Bool, 2}: a two-dimensional Boolean array.

Returns

Array{Bool, 2}: a two-dimensional Boolean array of the same size as the input, where true indicates the presence of an edge and false indicates its absence.

Notes

This function assumes that the input array is padded with false (or 0) at the boundaries. You might want to add boundary condition checking or padding if your input does not meet this condition.

This function uses "edge wrapping" when it shifts the array, meaning the values at the edge of the array "wrap around" to the opposite edge. If you want a different boundary condition (such as padding with false), you would need to adjust the creation of the shifted arrays accordingly.

Skylight.emissivity_profile — Method

Assumes unit energy in the rest frame of the emitter

Skylight.energies_quotients — Method

Assuming all photons are emitted with unit initial energy

Skylight.equatorial_ring_areas — Method

equatorial_ring_areas(edges, spacetime)

Approximate areas of the equatorial rings delimited by edges in spacetime. The spacetime must be stationary and axisymmetric. The areas are computed using the formula 2π*sqrt(g[2,2]*g[4,4])*Δr, where g is the metric evaluated at the center of the ring.

Skylight.event_horizon_radius — Method

event_horizon_radius(spacetime)

Radius of the event horizon of the given spacetime. Defined only for black hole spacetimes.

Skylight.horizons — Method

May return negative roots

Skylight.infer_num_bins — Method

infer_num_bins(q, at_source, start, stop, bin_size_conditioner, edge_width, camera)

Infer the number of bins from the energy quotients and the bin size conditioner. It chooses the larges number of bins that satisfies that the bin size is larger than the conditioner times the maximum local variation of the energy quotients, for a given approximation of the local variation.

Arguments

q: Array of energy quotients.

at_source: Array of booleans indicating whether the final position is at the source.

start: Lower bound of the range to be binned.

stop: Upper bound of the range to be binned.

bin_size_conditioner: Conditioner of the bin size.

edge_width: Width of the edge to be ignored for bin size conditioning (since edges have unusually large local variations, especially in the presence of light-rings and similar phenomena).

camera: Image plane.

Returns

num_bins: Number of bins.

Skylight.initialize — Method

Sets the bundle of rays at the camera position. The rays are initialized on a tetrad which has a static four-velocity and a spherical-like spatial triad. The rays have unit energy in this tetrad. See docs/pinholecamera.md for more details

Skylight.instantiate_custom_type — Method

instantiate_custom_type(dict)

Create an instance of a custom type using a dictionary. This version supports custom types created using the @with_kw macro from the Parameters.jl package and handles nested dictionaries. The dictionary and its nested subdictionaries must contain a _typename key whose value is the name of the custom type.

Arguments

dict::Dict{Symbol, }: A dictionary containing the custom type's name (stored in the _typename key) and field values.

Returns

An instance of the custom type with the field values specified in the input dictionary, including instances created from nested dictionaries.

Example

```julia using Parameters

@with_kw struct MyNestedType a::Int end

@with_kw struct MyTypeWithKW b::Float64 nested::MyNestedType end

mydict = Dict(:typename => "MyTypeWithKW", :b => 2.0, :nested => Dict(:_typename => "MyNestedType", :a => 1))

myinstance = instantiatecustomtype(mydict)

Skylight.inv4x4sym! — Method

Computes the inverse of a 4x4 symmetric matrix using the closed-form solution for the inverse.

Arguments:

inv_matrix: mutable array of size (4,4) to store the resulting inverse matrix.

matrix: array of size (4,4) containing the input symmetric matrix.

Returns: nothing.

Skylight.is_final_position_at_edge — Method

is_final_position_at_edge(output_data, configurations)

Determine whether the final position of each ray is at the edge of the source for a given set of configurations.

Arguments

output_data: a multi-dimensional array containing output data for each ray.

configurations: a data structure containing the spacetime and radiative model configurations.

Returns

A boolean grid indicating whether each ray's final position is at the edge of the source.

Skylight.is_final_position_at_edge — Method

is_final_position_at_edge(width::Int, output_data, configurations)

Determine whether the final position of each ray is at the edge of the source for a given set of configurations, considering a specified edge width.

Arguments

width: the width (in number of cells) to consider as the edge of the source.

output_data: a multi-dimensional array containing output data for each ray.

configurations: a data structure containing the spacetime and radiative model configurations.

Returns

A boolean grid indicating whether each ray's final position is within the specified width of the edge of the source.

Skylight.is_final_position_at_observer — Method

is_final_position_at_observer(output_data, configurations)

Determine whether the final position of each ray is at the observer for a given set of configurations.

Arguments

output_data: a multi-dimensional array containing output data for each ray.

configurations: a data structure containing the spacetime and radiative model configurations.

Returns

A boolean grid indicating whether each ray's final position is at the observer.

Notes

The function is_final_position_at_observer should already be defined elsewhere, and is used within this function to check the final position of each ray.

Skylight.is_final_position_at_source — Method

is_final_position_at_source(output_data, configurations)

Determine whether the final position of each ray is at the source for a given set of configurations.

Arguments

output_data: a multi-dimensional array containing output data for each ray.

configurations: a data structure containing the spacetime and radiative model configurations.

Returns

A boolean grid indicating whether each ray's final position is at the source.

Notes

The function is_final_position_at_source should already be defined elsewhere, and is used within this function to check the final position of each ray.

Skylight.is_hdf5_supported_type — Method

is_hdf5_supported_type(value)

Check if a given value is of a type that can be directly saved using the HDF5.jl package. This function can be used to determine whether a custom type or one of its fields can be saved to an HDF5 file without further conversion.

According to the HDF5.jl documentation, the supported types include:

Signed and unsigned integers (8, 16, 32, and 64 bits)

Float32 and Float64

Complex versions of numeric types

Arrays of supported numeric types (including complex versions)

AbstractString (ASCIIString and UTF8String)

Arrays of supported string types (ASCIIString and UTF8String)

Arguments

value: The value to be checked for HDF5 support.

Returns

A Bool indicating whether the input value is of a type supported by HDF5.jl.

Skylight.isco_radius — Method

isco_radius(spacetime, rotation_sense::AbstractRotationSense)

Radius of the innermost stable circular orbit in the given spacetime and rotation sense (prograde or retrograde), if defined.

Skylight.isco_radius — Method

Set for M=1

Skylight.line_emission_spectrum — Method

line_emission_spectrum(initial_data, output_data, configurations::VacuumOTEConfigurations; 
                       bin_size::Number=NaN, num_bins::Int=NaN,
                       start::Number=NaN, stop::Number=NaN)

Compute the binned intensity of a line emission spectrum.

Arguments

initial_data: Initial condition data.

output_data: Output data from the radiation model.

configurations::VacuumOTEConfigurations: Configuration parameters for the model.

Keywords

bin_size::Union{Nothing, Number}=nothing: Size of each bin. Either bin_size or num_bins must be specified.

num_bins::Union{Nothing, Number}=nothing: Number of bins. Either bin_size or num_bins must be specified.

start::Union{Nothing, Number}=nothing: Lower bound of the range to be binned. If unspecified, the minimum of the energy quotients will be used.

stop::Union{Nothing, Number}=nothing: Upper bound of the range to be binned. If unspecified, the maximum of the energy quotients will be used.

observer_four_velocity::AbstractVector (optional): The four-velocity of the observer. If not provided, a default static four-velocity will be used.

flux_direction::AbstractVector (optional): The direction in which to measure the flu. If not provided, a default direction will be used.

Returns

binned_fluxes: Array of the binned intensity in each bin.

bins: Array of the bin edges.

Notes

observer_four_velocity and flux_direction are only accepted if configurations.camera is a PinholeCamera.

Skylight.load_callback_params_from_hdf5 — Method

load_callback_params_from_hdf5(filename::String, run_index::Int)

Load the callback parameters of a specific run from an HDF5 file and instantiate the custom type if possible.

Arguments

filename: The path to the HDF5 file.

run_index: The index of the run to load the callback parameters from.

Returns

An instance of the custom type representing the callback parameters, or a dictionary if the custom type cannot be instantiated.

Skylight.load_configurations_from_hdf5 — Method

load_configurations_from_hdf5(filename::String)

Load configurations data from an HDF5 file.

Arguments

filename::String: The path to the HDF5 file.

Returns

An instance of the configuration type containing the configurations data stored in the HDF5 file.

Skylight.load_initial_data_from_hdf5 — Method

load_initial_data_from_hdf5(filename::String)

Load initial data from an HDF5 file.

Arguments

filename::String: The path to the HDF5 file.

Returns

The initial data stored in the HDF5 file.

Skylight.load_nested_dict_from_hdf5 — Method

load_nested_dict_from_hdf5(group::HDF5Group)

Convert an HDF5 group to a dictionary. If the value is "nothing", the value in the dict will be nothing.

Arguments

group::HDF5Group: The HDF5 group to convert to a dictionary.

Returns

A dictionary containing the data stored in the HDF5 group.

Skylight.load_output_data_from_hdf5 — Method

load_output_data_from_hdf5(filename::String, run_index::Int)

Load the output data of a specific run from an HDF5 file.

Arguments

filename: The path to the HDF5 file.

run_index: The index of the run to load the output data from.

Returns

The output data of the specified run.

Skylight.load_output_data_from_hdf5 — Method

load_output_data_from_hdf5(filename::String, run_indices::Vector{Int})

Load output data for a selected set of runs from an HDF5 file.

Arguments

filename::String: The path to the HDF5 file.

run_indices::Vector{Int}: A vector of integers indicating which runs' output data to load.

Returns

An array containing the output data for the selected runs.

Skylight.load_runs_from_hdf5 — Method

load_runs(filename::String, run_indices::Vector{Int})

Load a set of runs specified by their indices from an HDF5 file.

Arguments

filename: The path to the HDF5 file.

run_indices: A vector containing the indices of the runs to be loaded.

Returns

A vector of tuples, each containing the outputdata, callback dictionary, callbackparameters, and kwargs dictionary for the specified runs.

Skylight.lorentz_factors — Method

lorentz_factors(positions, spacetime, model)

Lorentz factors of the rest frame four velocities of `model` at given `positions`.

Skylight.mbco_radius — Method

mbco_radius(spacetime, rotation_sense::AbstractRotationSense)

Radius of the marginally bound circular orbit in the given spacetime and rotation sense (prograde or retrograde), if defined.

Skylight.mbco_radius — Method

Set for M=1

Skylight.metric! — Method

metric!(g, position, spacetime, cache)

Evaluate the spacetime metric at the given position and store the result in the given array using cache for temporary storage.

Skylight.metric! — Method

"Ruffini-Arguelles-Rueda metric for dark-matter halo

Skylight.metric — Method

metric(position, spacetime, cache)

Evaluate the spacetime metric at the given position and return the result using a cache object as temporary storage.

Skylight.metric — Method

metric(position, spacetime)

Evaluate the spacetime metric at the given position and return the result.

Skylight.metric_inverse! — Method

metric_inverse!(ginv, position, spacetime, g, cache)

Evaluate the inverse of the metric at the given position and store the result in the given array, using g and cache for temporary storage. Unless a more specialized method is defined, a fast inversion method for 4x4 symmetric matrices is used.

Skylight.metric_jacobian! — Method

metric_jacobian!(∂g, position, spacetime::AbstractSpacetime, g)

Computes the Jacobian matrix of the metric function with respect to spacetime coordinates using forward-mode automatic differentiation.

Arguments:

∂g: mutable array of size (4,4,4) to store the resulting Jacobian matrix.

position: tuple of four numbers representing a position in spacetime.

spacetime: object representing the spacetime.

g: array of size (4,4) containing the metric evaluated at the given position.

Returns: nothing.

Skylight.metric_jacobian! — Method

metric_jacobian!(∂g, position, spacetime_metric_field::Function, g, cfg::ForwardDiff.JacobianConfig)

Computes the Jacobian matrix of the metric function with respect to spacetime coordinates using forward-mode automatic differentiation.

Arguments:

∂g: mutable array of size (4,4,4) to store the resulting Jacobian matrix.

position: tuple of four numbers representing a position in spacetime.

spacetime: object representing the spacetime.

g: array of size (4,4) containing the metric evaluated at the given position.

cfg: object of type ForwardDiff.JacobianConfig with preallocated work buffers

Returns: nothing.

Skylight.midpoints — Method

midpoints(edges)

Returns midpoints of given edges

Skylight.observed_bolometric_intensities — Method

observedbolometricintensities(initialdata, outputdata, configurations::VacuumOTEConfigurations, camera::PinholeCamera; observerfourvelocity=nothing)

Compute observed bolometric intensities and energy quotients for a set of rays defined by the initial and final conditions. The function also checks whether the final position of each ray is at the source, in which case it proceeds with computation, else it skips to the next ray and the values are set to zero.

# Arguments

initial_data::AbstractMatrix: A matrix containing the initial data of the rays. The first four entries for each ray represent the initial position, while entries five to eight represent the initial momentum.

output_data::AbstractMatrix: A matrix containing the final data of the rays. The first four entries for each ray represent the final position, while entries five to eight represent the final momentum.

configurations::VacuumOTEConfigurations: The configurations with which the initial and output data were obtained.

camera::PinholeCamera: The camera model used to capture the scene.

observer_four_velocity::AbstractVector (optional): The four-velocity of the observer. If not provided, a default static four-velocity will be used.

Returns

Iobs::Vector: A vector of observed bolometric intensities for each ray, normalized by the distance to the image plane. The intensities are in CGS units.

q::Vector: A vector of energy quotients for each ray, representing the ratio of final energy in the local emission frame to initial energy in the observer frame.

Notes

Output units are CGS. The observer four-velocity and flux direction, if provided, must satisfy the conditions of being timelike and spacelike, respectively, as per the observer metric.

Skylight.observed_specific_intensities — Method

observed_specific_intensities(initial_data, output_data, configurations::VacuumOTEConfigurations)

Compute observed specific intensities and energy quotients for a set of rays defined by the initial and final conditions. The function also checks whether the final position of each ray is at the source, in which case it proceeds with computation, else it skips to the next ray and the values are set to zero.

Arguments

initial_data::AbstractMatrix: A matrix containing the initial data of the rays. The first four entries for each ray represent the initial position, while entries five to eight represent the initial momentum.

output_data::AbstractMatrix: A matrix containing the final data of the rays. The first four entries for each ray represent the final position, while entries five to eight represent the final momentum.

configurations::VacuumOTEConfigurations: The configurations with which the initial and output data were obtained.

observation_energies::Vector: A vector of energies in CGS units at which the specific intensities are to be computed.

Returns

Iobs::Vector: A vector of observed specific intensities in CGS units for each ray, normalized by the distance to the image plane. The intensities are in CGS units.

q::Vector: A vector of energy quotients for each ray, representing the ratio of final energy in the local emisison frame to initial energy in the observer frame.

Notes

Input energy units must be CGS. Output units are CGS.

Skylight.observed_specific_intensities — Method

observed_specific_intensities(initial_data, output_data, configurations::VacuumOTEConfigurations, observation_energies; observer_four_velocity=nothing)

Compute observed specific intensities and energy quotients for a set of rays defined by the initial and final conditions. The function also checks whether the final position of each ray is at the source, in which case it proceeds with computation, else it skips to the next ray and the values are set to zero.

Arguments

initial_data::AbstractMatrix: A matrix containing the initial data of the rays. The first four entries for each ray represent the initial position, while entries five to eight represent the initial momentum.

output_data::AbstractMatrix: A matrix containing the final data of the rays. The first four entries for each ray represent the final position, while entries five to eight represent the final momentum.

configurations::VacuumOTEConfigurations: The configurations with which the initial and output data were obtained.

observation_energies::Vector: A vector of energies in CGS units at which the specific intensities are to be computed.

observer_four_velocity::AbstractVector (optional): The four-velocity of the observer. If not provided, a default static four-velocity will be used.

Returns

Iobs::Matrix: A matrix of observed specific intensities in CGS units for each ray, normalized by the distance to the image plane. The intensities are in CGS units.

q::Vector: A vector of energy quotients for each ray, representing the ratio of final energy in the local emission frame to initial energy in the observer frame.

Notes

Input energy units must be CGS. Output units are CGS. The observer four-velocity and flux direction, if provided, must satisfy the conditions of being timelike and spacelike, respectively, as per the observer metric.

Skylight.orthogonal_projection — Method

Projection of v orthogonal to normalized timelike vector u

Skylight.orthonormalize! — Method

orthonormalize!(dyad::Matrix, time_vector::Vector, space_vector::Vector, metric)

Orthonormalize the dyad `dyad` with respect to the metric `metric` and the vectors `time_vector` and `space_vector`.
Both `time_vector` and `space_vector` are assumed to be normalized and orthogonal to each other.

Skylight.orthonormalize! — Method

orthonormalize!(triad::Matrix, time_vector::Vector, metric)

Orthonormalize the triad `triad` with respect to the metric `metric` and the vector `time_vector`.
`time_vector` is assumed to be normalized.

Skylight.outermost_circular_orbit_radius — Method

Just return 100M, this if for ion torus

Skylight.packets_tetrad_components! — Method

By having Minkowski-null components in the tetrad we guarantee that the momentum is null

Skylight.packets_unit_random_triad_components! — Method

Sets only positive components along the surface normal

Skylight.planck_function — Method

planck_function(E, T)

Planck's law of black body radiation in terms of energy.

Arguments

E::Float64: The energy in erg.

T::Float64: The temperature in Kelvin.

Returns

::Float64: The planck function in CGS.

Skylight.planck_integral — Method

planck_integral(E, T)

Estimation of the Planck integral from E to infinity for temperature T. Based on "Planck functions and integrals; methods of computation" (T.E. Michels) https://ntrs.nasa.gov/archive/nasa/casi.ntrs.nasa.gov/19680008986.pdf

Arguments

E::Float64: The energy in erg.

T::Float64: The temperature in Kelvin.

Returns

::Float64: The planck integral in CGS.

Remarks

An evaluation of the error is too expensive for millions of calls, so we fix N=5. This was checked against scipy integration, yielding at least 3 siginficant digits for all x. When x>1 the amount of siginificant digits is 8 or more

Skylight.points_on_equatorial_plane! — Method

points_on_equatorial_plane!(points, r, φ, ::CartesianTopology)

Sets the points on the equatorial plane from spherical coordinates

Skylight.points_on_unit_sphere! — Method

points_on_unit_sphere!(points, θ, φ, ::AbstractCoordinatesTopology)

Transforms the points on the unit sphere from the angular coordinates

Skylight.radius — Method

radius(position, spacetime)

Radius of the given position in the given spacetime.

Skylight.radius — Method

radius(position, spacetime::SchwarzschildSpacetimeKerrSchildCoordinates)

``r = \sqrt{x^2 + y^2 + z^2}``

Skylight.random_azimuthal_angle — Method

random_azimuthal_angle(N)

Generates N random numbers representing azimuthal angles, following a uniform distribution from 0 to 2π.

Skylight.random_cylindrical_radius — Method

random_cylindrical_radius(N, rmin, rmax)

Generates N random numbers representing cylindrical radii, following a uniform distribution between rmin and rmax.

Skylight.random_polar_angle — Method

random_polar_angle(N, θmax)

Generates N random numbers representing polar angles, following a uniform distribution from 0 to θmax (in radians).

Skylight.random_uniform_points_annulus! — Method

random_uniform_points_annulus!(points, rmin, rmax, coords_top)

Populate the given array points with points randomly and uniformly distributed in an annulus with inner radius rmin and outer radius rmax. The type of the coordinate system is determined by coords_top.

Skylight.random_uniform_points_disk! — Method

random_uniform_points_disk!(points, rmax, coords_top)

Populate the given array points with points randomly and uniformly distributed in a disk with radius rmax. The type of the coordinate system is determined by coords_top.

Skylight.random_uniform_points_unit_hemisphere! — Method

random_uniform_points_unit_hemisphere!(points, coords_top)

Populate the given array points with points randomly and uniformly distributed on the unit hemisphere. The type of the coordinate system is determined by coords_top.

Skylight.random_uniform_points_unit_hemisphere_xaxis! — Method

random_uniform_points_unit_hemisphere_xaxis!(v,::CartesianTopology)

Populate the given array v with points randomly and uniformly distributed on the unit hemisphere along the x-axis.

Skylight.random_uniform_points_unit_sphere! — Method

random_uniform_points_unit_sphere!(points, coords_top)

Populate the given array points with points randomly and uniformly distributed on the unit sphere. The type of the coordinate system is determined by coords_top.

Skylight.random_uniform_points_unit_spherical_cap! — Method

random_uniform_points_unit_spherical_cap!(points, θmax_in_degrees, coords_top)

Populate the given array points with points randomly and uniformly distributed in a spherical cap with maximum polar angle θmax_in_degrees degrees. The type of the coordinate system is determined by coords_top.

Skylight.rayleigh_jeans_function — Method

rayleigh_jeans_function(E, T)

Rayleigh-Jeans law of black body radiation in terms of energy.

Arguments

E::Float64: The energy in erg.

T::Float64: The temperature in Kelvin.

Returns

::Float64: The Rayleigh-Jeans function in CGS.

Skylight.rays_tetrad_components! — Method

By having Minkowski-null components in the tetrad we guarantee that the momentum is null

Skylight.save_nested_dict_to_hdf5 — Method

save_nested_dict_to_hdf5(group::HDF5File, nested_dict::Dict{Symbol,})

Recursively save a nested dictionary to an HDF5 group. The function iterates through the key-value pairs in the input dictionary. If a value is a dictionary, it creates a new subgroup and recursively saves the nested dictionary. If a value is not a dictionary, it is saved directly to the current group.

Arguments

group: An HDF5 group to which the nested dictionary should be saved.

nested_dict: A nested dictionary with keys of type Symbol and values of any type.

Skylight.save_obj_to_hdf5 — Method

save_obj_to_hdf5(group::HDF5File, name::String, obj::Any)

Save an object to an HDF5 group by converting it to an HDF5 compatible dictionary and then saving the dictionary to the group. The function creates a new subgroup with the given name and saves the converted dictionary using the save_nested_dict_to_hdf5 function.

Arguments

group: An HDF5 group to which the object should be saved.

name: A string representing the name of the subgroup to create for the object.

obj: An object of any type that should be saved to the HDF5 group.

Skylight.save_to_hdf5 — Method

save_to_hdf5(filename, configurations, initial_data, runs)

Save the provided data to an HDF5 file. The data includes configurations, initialdata, and a collection of runs. Each run contains outputdata, callback, callback_parameters, and kwargs for the solver function.

If the file exists, it uses the "r+" mode to open the file for reading and writing. If the file does not exist, it uses the "w" mode to create a new file.

The file will be organized as follows:

/configurations: A group containing the configurations data.

/initial_data: A dataset containing the initial data.

/runs: A group containing subgroups for each run. Each subgroup will have:

/output_data: A dataset containing the output data.

/callback: A subgroup containing the callback custom type data.

/callback_parameters: A subgroup containing the callback parameters data.

/kwargs: A subgroup containing the kwargs data.

Arguments

filename: The name of the HDF5 file to save the data to. If the file already exists, new data will be appended without overwriting existing content.

configurations: A custom type containing the configurations data. It will be converted to a dictionary using the to_hdf5_compatible_dict function.

initial_data: The initial data for the equations.

runs: An array of tuples containing outputdata, callback custom type, callbackparameters, and kwargs for each run. The callback custom type and callbackparameters should be custom types, which will be converted to dictionaries using the `tohdf5compatibledict` function.

Returns

Nothing.

Skylight.set_null! — Method

returns with unit energy

Skylight.set_null_ingoing_past_directed! — Method

the input time component must be zero for this to work

Skylight.spherical_symmetry — Method

spherical_symmetry(spacetime)

Return IsSphericallySymmetric() if the spacetime is spherically symmetric, IsNotSphericallySymmetric() otherwise.

Skylight.stationarity — Method

stationarity(spacetime)

Return IsStationary() if the spacetime is stationary, IsNotStationary() otherwise.

Skylight.tmap — Method

tmap(f, itr; tasks_per_thread::Int = 2, kwargs...)

Apply a given function f to partitions of the iterable itr in parallel.

Arguments

f: A function to be applied to each partition of the iterable.

itr: The iterable to be partitioned and processed.

tasks_per_thread::Int: An optional parameter to control the number of tasks per thread. Default value is 2.

kwargs...: Additional keyword arguments that will be passed to the function f.

Description

This function divides the iterable itr into chunks, with each chunk size determined by the given tasks_per_thread and the number of available threads. The specified function f is then applied to each chunk in parallel using the @spawn macro. The @sync macro ensures that all tasks are completed before moving on.

Example

```julia function printelements(chunk) for i in chunk println(i) end end tmap(printelements, 1:10; tasksperthread = 2)

Skylight.to_hdf5_compatible_dict — Method

to_hdf5_compatible_dict(obj::T; depth::Int=0, max_depth::Int=5) where T

Recursively convert a custom type obj to a dictionary. This function is useful for preparing custom types for saving to an HDF5 file when the types are not directly supported by the HDF5.jl package.

The function will attempt to convert each field of the custom type to a dictionary if the field is not an HDF5-supported type. The conversion process is limited by a maximum recursion depth, specified by the max_depth parameter, which defaults to 5.

If a field is nothing, the value will be saved as the string "nothing".

If a field is a function, the function will be converted to a string using the string function.

If a field is not an HDF5-supported type and has no fields that can be converted, the function will return nothing when the maximum recursion depth is reached.

Arguments

obj::T: The custom type instance to be converted to a dictionary.

depth::Int=0 (optional): The current recursion depth, starting at 0. This parameter is used internally and should not typically be set by the user.

max_depth::Int=5 (optional): The maximum recursion depth allowed for converting nested custom types.

Returns

A Dict{String, Any} containing the fields of the input custom type, with nested custom types recursively converted to dictionaries. The dictionary will also include a "_typename" key with the name of the custom type as a string.

Skylight.to_hdf5_compatible_dict — Method

to_hdf5_compatible_dict(cb::T; depth::Int=0, max_depth::Int=8) where {T<:SciMLBase.DECallback}

Converts a DifferentialEquations.jl callback into a dictionary that can be saved to an HDF5 file for later instantiation.

Arguments

cb: The callback to be converted into a dictionary.

depth: The current depth of recursion. This is used to prevent infinite recursion. (default: 0)

max_depth: The maximum allowed depth of recursion. If the depth exceeds this value, the function returns nothing. (default: 8)

Returns

A dictionary where the keys are the names of the fields of the object and the values are the corresponding field values. Field values are recursively converted to dictionaries if they are of custom types not supported by HDF5.jl.

Notes

The dictionary includes an additional entry with the key "_typename" and the name of the type of the object as the value. This is used to reconstruct the original object when loading the data from the HDF5 file.

The function specifically handles certain fields of the DECallback type.

Skylight.volume_element — Method

volume_element(position, spacetime, g, cache)

Compute the volume element (square root of minus the determinant of the metric) at the given position with a fast determinant for 4x4 symmetric matrices, using g and cache for temporary storage.

Skylight.volume_element — Method

volume_element(position, spacetime)

Compute the volume element (square root of minus the determinant of the metric) at the given position with a fast determinant for 4x4 symmetric matrices.