Other types

Skylight.AbstractBlackHoleSpacetime — Type

AbstractBlackHoleSpacetime <: AbstractSpacetime

Supertype for representing spacetimes containing black holes.

Skylight.AbstractChristoffelCache — Type

AbstractSpacetimeCache

Abstract type for cache objects to be used as temporary storage in Christoffel symbol calculations.

Skylight.AbstractCoordinatesTopology — Type

AbstractCoordinatesTopology

Abstract type for representing the topology of the coordinates of a spacetime.

Skylight.AbstractRegularCompactObjectSpacetime — Type

AbstractRegularCompactObjectSpacetime <: AbstractSpacetime

Supertype for representing spacetimes containing compact objects without horizons like boson stars.

Skylight.AbstractRotationSense — Type

AbstractRotationSense

Abstract type for representing the rotation sense of a circular orbit in a spacetime.

Skylight.AbstractSpacetime — Type

AbstractSpacetime

Supertype for representing the geometrical structure of spacetime. Specific implementations of spacetime, such as black holes and regular compact objects, are subtypes of this abstract type.

Skylight.AbstractSpacetimeCache — Type

AbstractSpacetimeCache

Abstract type for caching spacetime-related computations. This can be used as scratch memory in calculations involving the spacetime.

Skylight.CartesianTopology — Type

CartesianTopology <: AbstractCoordinatesTopology

Cartesian coordinates topology.

Skylight.ProgradeRotation — Type

ProgradeRotation <: AbstractRotationSense

Prograde rotation sense for a circular orbit in a spacetime.

Constructor

ProgradeRotation()

Skylight.RetrogradeRotation — Type

RetrogradeRotation <: AbstractRotationSense

Retrograde rotation sense for a circular orbit in a spacetime.

Constructor

RetrogradeRotation()

Skylight.SphericalTopology — Type

SphericalTopology <: AbstractCoordinatesTopology

Spherical coordinates topology.