Spacetimes

Skylight.AbstractSpacetime — Type

AbstractSpacetime

Supertype for representing the geometrical structure of spacetime. Specific implementations of spacetime, such as black holes and regular compact objects, are subtypes of this abstract type.

Skylight.AbstractBlackHoleSpacetime — Type

AbstractBlackHoleSpacetime <: AbstractSpacetime

Supertype for representing spacetimes containing black holes.

Skylight.AbstractRegularCompactObjectSpacetime — Type

AbstractRegularCompactObjectSpacetime <: AbstractSpacetime

Supertype for representing spacetimes containing compact objects without horizons like boson stars.

Skylight.MinkowskiSpacetimeCartesianCoordinates — Type

MinkowskiSpacetimeCartesianCoordinates <: AbstractSpacetime

Minkowski spacetime using Cartesian coordinates. The metric is

$ds^2 = -dt^2 + dx^2 + dy^2 + dz^2$

See also: Minkowski Spacetime.

Skylight.MinkowskiSpacetimeSphericalCoordinates — Type

MinkowskiSpacetimeSphericalCoordinates <: AbstractSpacetime

Minkowski spacetime in spherical coordinates. The metric is

$ds^2 = -dt^2 + dr^2 + r^2 d\theta^2 + r^2 \sin^2 \theta d\phi^2$

where $r$ is the radial coordinate and $θ$ is the polar angle.

See also: Minkowski Spacetime.