Catalogue of spacetimes

Minkowski spacetime

Skylight.MinkowskiSpacetimeCartesianCoordinates — Type

MinkowskiSpacetimeCartesianCoordinates <: AbstractMinkowskiSpacetime

Minkowski Spacetime in spherical coordinates.

$ds^2 = -dt^2 + dx^2 + dy^2 + dz^2$

Constructor

MinkowskiSpacetimeCartesianCoordinates()

Skylight.MinkowskiSpacetimeSphericalCoordinates — Type

MinkowskiSpacetimeSphericalCoordinates <: AbstractMinkowskiSpacetime

Minkowski Spacetime in spherical coordinates.

$ds^2 = -dt^2 + dr^2 + r^2 d\theta^2 + r^2 \sin^2 \theta d\phi^2$

where $r$ is the radial coordinate and $θ$ is the polar angle.

Constructor

MinkowskiSpacetimeSphericalCoordinates()

Schwarzschild spacetime

Skylight.SchwarzschildSpacetimeKerrSchildCoordinates — Type

SchwarzschildSpacetimeKerrSchildCoordinates <: AbstractSchwarzschildSpacetime

Schwarzschild spacetime in Kerr-Schild coordinates. The parameter $M$ is the mass. The metric is

$g_{\mu \nu} = \eta_{\mu \nu} + H l_{\mu} l_{\nu}$

where $\eta_{\mu \nu}$ is the flat metric, $H=2M/r$, and $l_{\mu}=(1,x,y,z)/r$.

Constructor

SchwarzschildSpacetimeKerrSchildCoordinates(M=1.0)

Skylight.SchwarzschildSpacetimeSphericalCoordinates — Type

SchwarzschildSpacetimeSphericalCoordinates <: AbstractSchwarzschildSpacetime

Schwarzschild spacetime in spherical coordinates. The metric is

$ds^2 = -(1-2M/r) dt^2 + (1-2M/r)^{-1} dr^2 + r^2 d\theta^2 + r^2 \sin^2 \theta d\phi^2$

Constructor

SchwarzschildSpacetimeSphericalCoordinates(M=1.0)

Kerr spacetime

Skylight.KerrSpacetimeBoyerLindquistCoordinates — Type

KerrSpacetimeBoyerLindquistCoordinates <: AbstractKerrSpacetime

Kerr spacetime in Boyer-Lindquist coordinates. The parameter are the mass $M$ and the dimensionless spin $a$.

Constructor

KerrSpacetimeBoyerLindquistCoordinates(M=1.0, a=0.99)

Skylight.KerrSpacetimeKerrSchildCoordinates — Type

KerrSpacetimeKerrSchildCoordinates <: AbstractKerrSpacetime

Kerr spacetime in Kerr-Schild coordinates. The parameter are the mass $M$ and the dimensionless spin $a$. The metric is

$g_{\mu \nu} = \eta_{\mu \nu} + H l_{\mu} l_{\nu}$

where $\eta_{\mu \nu}$ is the flat metric, $H=2Mr^3/(r^4+a^2 z^2$, and

$l_{\mu}=(1,\frac{rx+ay}{r^2+a^2},\frac{ry-ax}{r^2+a^2},z/r)$

where $r$ satisfies the equation

$\frac{x^2+y^2}{r^2+a^2} + \frac{z^2}{r^2} = 1$

Constructor

KerrSpacetimeKerrSchildCoordinates(M=1.0, a=0.99)

Johannsen spacetime

Skylight.JohannsenSpacetime — Type

JohannsenSpacetime <: AbstractBlackHoleSpacetime

Johannsen spacetime to lowest order in the deformation parameters with respect to the Kerr metric

Constructor

JohannsenSpacetime(M=1.0, a=0.99, α13=0.5, α22=0.0, α52=0.01, ϵ3=0.0)

f(R)-Kerr spacetime

Skylight.FRKerrSpacetime — Type

FRKerrSpacetime <: AbstractBlackHoleSpacetime

$f(R)$-Kerr spacetime (uncharged) with constant curvature $R_0$

Constructor

FRKerrSpacetime(M=1.0, a=0.99, R0=-0.0012) 

Charged wormhole spacetime

Skylight.ChargedWormholeSpacetimeRegularCoordinates — Type

ChargedWormholeSpacetimeRegularCoordinates <: AbstractSpacetime

Charged wormhole spacetime in regular (throat crossing) coordinates.

Constructor

ChargedWormholeSpacetimeRegularCoordinates(b0=1.0, Q=0.5)

Skylight.ChargedWormholeSpacetimeSphericalCoordinates — Type

ChargedWormholeSpacetimeSphericalCoordinates <: AbstractSpacetime

Charged wormhole spacetime in usual spherical coordinates.

Constructor

ChargedWormholeSpacetimeSphericalCoordinates(b0=1.0, Q=0.5)

RAR spacetime

Skylight.RARSpacetime — Type

RARSpacetime <: AbstractRegularCompactObjectSpacetime

Ruffini-Arguelles-Rueda spacetime for dark-matter galactic core-halo. The spacetime is constructed by interpolating the data files in a given directory. The numerical data must include the $r$, $g_{tt}$, $g_{rr}$, $\partial_r \nu$, $M(r)$ and $\partial_r M$ as columns in that order. The data must be in geometrized units.

Constructor

RARSpacetime("./rar_data")

Boson star spacetime

Skylight.BosonStarSpacetime — Type

BosonStarSpacetime <: AbstractRegularCompactObjectSpacetime

Boson star spacetime in spherical coordinates. It uses analyical fits. Either the fit parameters as described in... can be provided as vectors, or any of the symbols :LBS1, :LBS2, :LBS3, :SBS1, :SBS2 or :SBS3 as constructor arguments

Constructors

BosonStarSpacetime(a=aparams, b=bparams) 
BosonStarSpacetime(:LBS1)