Geometric functions

Skylight.metric — Method

metric(position::AbstractVector, spacetime::AbstractSpacetime)

Evaluate the spacetime metric at the given position and return the result.

Skylight.metric_inverse — Method

metric_inverse(position::AbstractVector, spacetime::AbstractSpacetime)

Evaluate the inverse of the metric at the given position and return the result.

Skylight.volume_element — Method

volume_element(position::AbstractVector, spacetime::AbstractSpacetime)

Compute $\sqrt{-\text{det}(g)}$ at the given position

Skylight.christoffel — Method

christoffel(position::AbstractVector, spacetime::AbstractSpacetime)

Evaluate the Christoffel symbols of the second kind at the given position and return the result as a 4x4x4 array Γ, where Γ[α,μ,ν] corresponds to

$\Gamma^\alpha_{\mu \nu} = \frac{1}{2} g^{\alpha \rho}(\partial_\mu g_{\rho \nu} + \partial_\nu g_{\rho \mu} - \partial_\rho g_{\mu \nu}).$

Non-allocating methods

For non-allocating versions to be used within tight loops, use the following methods

Skylight.metric! — Method

metric!(g::AbstractMatrix, position::AbstractVector, spacetime::AbstractSpacetime, cache::Union{Nothing, AbstractSpacetimeCache})

Evaluate the spacetime metric at the given position and store the result in g using cache for temporary storage.

Skylight.metric_inverse! — Method

metric_inverse!(ginv::AbstractMatrix, position::AbstractVector, spacetime::AbstractSpacetime, g::AbstractMatrix, cache::Union{Nothing, AbstractSpacetimeCache})

Evaluate the inverse of the metric at the given position and store the result in the given array, using g and cache for temporary storage, which must be preallocated as g = zeros(4,4) and cache = allocate_cache(spacetime).

See also allocate_cache.

Skylight.volume_element — Method

volume_element(position::AbstractVector, spacetime::AbstractSpacetime, g::AbstractMatrix, cache::Union{Nothing, AbstractSpacetimeCache})

Compute $\sqrt{-\text{det}(g)}$ at the given position, using g and cache for temporary storage, which must be preallocated as g = zeros(4,4) and cache = allocate_cache(spacetime).

See also allocate_cache.

Skylight.christoffel! — Method

christoffel!(Γ::AbstractArray, position::AbstractVector, spacetime::AbstractSpacetime, cache::Union{Nothing,AbstractChristoffelCache})

Evaluate the Christoffel symbols of the second kind at the given position and store the result in Γ, where Γ[α,μ,ν] corresponds to

$\Gamma^\alpha_{\mu \nu} = \frac{1}{2} g^{\alpha \rho}(\partial_\mu g_{\rho \nu} + \partial_\nu g_{\rho \mu} - \partial_\rho g_{\mu \nu}).$

The cache must be preallocated as cache = allocate_christoffel_cache(spacetime).

See also allocate_christoffel_cache.

Cache allocation

Skylight.allocate_cache — Method

allocate_cache(spacetime::AbstractSpacetime)

Allocate a cache object for the given spacetime. The cache object is used to store temporary data in spacetime-related calculations.

Skylight.allocate_christoffel_cache — Method

allocate_christoffel_cache(spacetime::AbstractSpacetime)

Allocate a cache object for Christoffel symbols calculations.

See also christoffel!.